M3前最后一次提交

+++ build.cjs
@@ -6,7 +6,9 @@ const archiever = require('archiver')
 fs.copyFileSync(path.join(__dirname, 'src', 'graaljs.d.ts'), path.join(__dirname, 'lib', 'graaljs.d.ts'))
 fs.copyFileSync(path.join(__dirname, 'plugin.json'), path.join(__dirname, 'lib', 'plugin.json'))
 
-fs.unlinkSync(path.join(__dirname, 'jslib.zip'))
+if(fs.existsSync(path.join(__dirname, 'jslib.zip'))) {
+  fs.unlinkSync(path.join(__dirname, 'jslib.zip'))
+}
 const output = fs.createWriteStream(path.join(__dirname, 'jslib.zip'))
 const archive = archiever('zip')
 archive.pipe(output)
